% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.mm.to.hs.R
\name{ensembl.mm.to.hs}
\alias{ensembl.mm.to.hs}
\title{ensembl.mm.to.hs
this function take input vector of murine Ensembl identifiers and convert them
in human identifiers according Biomart database version 110}
\usage{
library(geneconverter)

mouse.ensembl<-c("ENSMUSG00000005672","ENSMUSG00000058488","ENSMUSG00000061353","ENSMUSG00000016494")

ensembl.mm.to.hs(mouse.ensembl)
}
\description{
ensembl.mm.to.hs
this function take input vector of murine Ensembl identifiers and convert them
in human identifiers according Biomart database version 110
}
\examples{
library(geneconverter)
mouse.ensembl<-c("ENSMUSG00000005672","ENSMUSG00000058488","ENSMUSG00000061353","ENSMUSG00000016494")
ensembl.mm.to.hs(mouse.ensembl)
}
