% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.hs.to.mm.R
\name{gene.hs.to.mm}
\alias{gene.hs.to.mm}
\title{gene.hs.to.mm
this function take input vector of human gene symbols and convert them
in mouse gene symbols according Biomart  database version 110}
\usage{
library(geneconverter)

human.gene<-c("FLT3","CXCR4","CXCR7","CD34","KIT","KL","CXCL12")

gene.hs.to.mm(human.gene)
}
\description{
gene.hs.to.mm
this function take input vector of human gene symbols and convert them
in mouse gene symbols according Biomart  database version 110
}
\examples{
library(geneconverter)
human.gene<-c("FLT3","CXCR4","CXCR7","CD34","KIT","KL","CXCL12")
gene.hs.to.mm(human.gene)
}
