% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.mm.to.hs.R
\name{gene.mm.to.hs}
\alias{gene.mm.to.hs}
\title{gene.mm.to.hs
this function take input vector of murine gene symbols and convert them
in human gene symbols according Biomart database version 110}
\usage{
library(geneconverter)

mouse.gene<-c("Kit","Sca","Kl","Cd34","Flt3","Cxcl12")

gene.mm.to.hs(mouse.gene)
}
\description{
gene.mm.to.hs
this function take input vector of murine gene symbols and convert them
in human gene symbols according Biomart database version 110
}
\examples{
library(geneconverter)
mouse.gene<-c("Kit","Sca","Kl","Cd34","Flt3","Cxcl12")
gene.mm.to.hs(mouse.gene)
}
