% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgncConverter.R
\name{hgncConverter}
\alias{hgncConverter}
\title{Convert human gene synonyms to HGNC approved gene names}
\usage{
hgncConverter(genelist, colname, drop = FALSE)
}
\arguments{
\item{genelist}{A dataframe object}

\item{colname}{Name of the column containing human gene synonyms to be converted to HGNC approved gene names}

\item{drop}{If it is TRUE, it will delete row or rows which there is no match between human gene synonyms and HGNC approved gene names}
}
\value{
data frame object
}
\description{
This function converts a column containing human gene synonyms within a data frame
to HGNC appreved gene names. If the column contains a mix of gene names and gene synonyms,
only gene synonyms will be converted to gene names and gene names will remain the same.
}
