% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbiConverter.R
\name{ncbiConverter}
\alias{ncbiConverter}
\title{Convert human gene synonyms to NCBI gene names}
\usage{
ncbiConverter(genelist, colname, drop = FALSE)
}
\arguments{
\item{genelist}{A data frame object}

\item{colname}{Name of the column containing human gene synonyms to be converted to human NCBI gene names}

\item{drop}{If it is TRUE, it will delete row or rows which there is no match between human gene synonyms and human NCBI gene names}
}
\value{
Data frame object
}
\description{
This function converts a column containing human gene synonyms within a data frame
to NCBI gene names. If the column contains a mix of gene names and gene synonyms,
only gene synonyms will be converted to gene names and gene names will remain the same.
}
