% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoConverter.R
\name{orthoConverter}
\alias{orthoConverter}
\title{Convert C. elegans, D. melanogaster, M. musculus, R. norvegicus, D. rerio and,
C. cerevisiae gene names to human gene names.}
\usage{
orthoConverter(genelist, colname, ctype, drop = FALSE)
}
\arguments{
\item{genelist}{A dataframe object}

\item{colname}{Name of the column containing gene synonyms to be converted to human gene names}

\item{ctype}{The type of creature whose gene names will be changed}

\item{drop}{If it is true, it will delete row or rows which there is no match between selected creature and human}
}
\description{
This function converts a column containing gene names of given organisms within a data frame
to human gene names. If the column contains a mix of these organisms and Human gene names,
only gene synonyms will be converted to gene names and gene names will remain the same.
}
