% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoMouse.R
\name{synoMouse}
\alias{synoMouse}
\title{Convert mouse gene synonyms to NCBI mouse gene names}
\usage{
synoMouse(genelist, colname, drop = FALSE)
}
\arguments{
\item{genelist}{A dataframe object}

\item{colname}{Name of the column containing mouse gene synonyms to be converted to mouse gene names}

\item{drop}{If it is TRUE, it will delete row or rows which there is no match between mouse gene synonyms and mouse gene names}
}
\value{
data frame object
}
\description{
This function converts a column containing mouse gene synonyms within a data frame
to mouse NCBI gene names. If the column contains a mix of mouse gene names and mouse gene synonyms,
only mouse gene synonyms will be converted to mouse gene names and mouse gene names will remain the same.
}
